Refactor route module name
